Add files via upload
Use IIF(replacing CASE with IIF) that allows to returns one of two values, depending on whether the Boolean expression evaluates to true or false.

Pivot table (query #3), create Year auxiliary field YEAR(o.[OrderDate]) AS [Year] created specifically for the purpose of ilolating data based on the year of Customers sales and subsequent aggregation. The values in this field will be used in pivoted data as colums name with aggregated results.